feat: Add `File::detect_newline_style()`, which does at it says. (#331)
