+++ git-config/src/file/access/read_only.rs
@@ -4,6 +4,7 @@ use bstr::BStr;
 use git_features::threading::OwnShared;
 
 use crate::file::{Metadata, MetadataFilter};
+use crate::parse::Event;
 use crate::{file, lookup, File};
 
 /// Read-only low-level access methods, as it requires generics for converting into
@@ -271,5 +272,27 @@ impl<'event> File<'event> {
     /// Return all events which are in front of the first of our sections, or `None` if there are none.
     pub fn frontmatter(&self) -> Option<impl Iterator<Item = &crate::parse::Event<'event>>> {
         (!self.frontmatter_events.is_empty()).then(|| self.frontmatter_events.iter())
+    }
+
+    /// Return the newline characters that have been detected in this config file or the default ones
+    /// for the current platform.
+    ///
+    /// Note that the first found newline is the one we use in the assumption of consistency.
+    pub fn detect_newline_style(&self) -> &BStr {
+        fn extract_newline<'a, 'b>(e: &'a Event<'b>) -> Option<&'a BStr> {
+            match e {
+                Event::Newline(b) => b.as_ref().into(),
+                _ => None,
+            }
+        }
+
+        self.frontmatter_events
+            .iter()
+            .find_map(extract_newline)
+            .or_else(|| {
+                self.sections()
+                    .find_map(|s| s.body.as_ref().iter().find_map(extract_newline))
+            })
+            .unwrap_or_else(|| if cfg!(windows) { "\r\n" } else { "\n" }.into())
     }
 }
